+++ crates/miden-objects/src/block/header.rs
@@ -2,9 +2,7 @@ use alloc::vec::Vec;
 
 use crate::{
     Digest, Felt, Hasher, ZERO,
-    account::AccountId,
     block::BlockNumber,
-    transaction::TransactionId,
     utils::serde::{ByteReader, ByteWriter, Deserializable, DeserializationError, Serializable},
 };
 
@@ -222,20 +220,6 @@ impl BlockHeader {
         elements.extend_from_slice(tx_kernel_commitment.as_elements());
         elements.extend_from_slice(proof_commitment.as_elements());
         elements.extend([block_num.into(), version.into(), timestamp.into(), ZERO]);
-        Hasher::hash_elements(&elements)
-    }
-
-    /// Computes a commitment to the provided list of transactions.
-    pub fn compute_tx_commitment(
-        updated_accounts: impl Iterator<Item = (TransactionId, AccountId)>,
-    ) -> Digest {
-        let mut elements = vec![];
-        for (transaction_id, account_id) in updated_accounts {
-            let [account_id_prefix, account_id_suffix] = <[Felt; 2]>::from(account_id);
-            elements.extend_from_slice(transaction_id.as_elements());
-            elements.extend_from_slice(&[account_id_prefix, account_id_suffix, ZERO, ZERO]);
-        }
-
         Hasher::hash_elements(&elements)
     }
 }