+++ CHANGELOG.md
@@ -4,9 +4,11 @@
 
 - [BREAKING] Add `TransactionHeader` and include it in batches and blocks (#1247).
 - [BREAKING] Hash keys in storage maps before insertion into the SMT (#1250).
+- Add `AccountTree` and `PartialAccountTree` wrappers and enforce ID prefix uniqueness (#1254).
+- Added a retry strategy for worker's health check (#1255).
 - Added getter for proof security level in `ProvenBatch` and `ProvenBlock` (#1259).
 - [BREAKING] Replaced the `ProvenBatch::new_unchecked` with the `ProvenBatch::new` method to initialize the struct with validations (#1260).
-- Added a retry strategy for worker's health check (#1255).
+- Added pretty print for `AccountCode` (#1273).
 - Added field identifier structs for component metadata (#1292).
 
 ## 0.8.1 (2025-03-26) - `miden-objects` and `miden-tx` crates only.